requirements.txt
pdfplumber>=0.10.0
anthropic>=0.39.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
openpyxl>=3.1.0
neo4j>=5.0.0
chromadb==0.4.24
beautifulsoup4>=4.12.0
requests>=2.31.0
pyvis>=0.3.0
